requirements.txt
ursina
xlib
websocket-client